feat: Add confirmation modal for expense deletion and success alerts
- Added SweetAlert2 modal to confirm expense deletion, ensuring user validation before removing an expense.
- Implemented success alerts using SweetAlert2 for successful creation and editing of expenses, enhancing user feedback.
- Minor UI adjustments to improve user experience during CRUD operations.

+++ src/presentation/components/ExpenseTable.tsx
@@ -6,6 +6,7 @@ import {
 import { useNavigate } from "react-router-dom";
 import { Expense } from "../../domain/models/Expense";
 import { useAppDispatch } from "../../application/hooks/useAppDispatch";
+import Swal from "sweetalert2";
 
 interface ExpenseSummaryProps {
   expenses: Expense[];
@@ -24,7 +25,26 @@ const ExpenseTable: React.FC<ExpenseSummaryProps> = ({
   };
 
   const handleDelete = (id: number) => {
-    dispatch(deleteExpenseThunk(id));
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "No podrás deshacer esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteExpenseThunk(id));
+        Swal.fire({
+          title: "Eliminado",
+          text: "El gasto ha sido eliminado exitosamente.",
+          icon: "success",
+          confirmButtonColor: "#3085d6",
+        });
+      }
+    });
   };
   return (
     <div className="table-responsive">
